Android. Продвинутый уровень. ПЗ #1 (окончательный рефакторинг)

+++ app/src/main/java/com/geekbrains/android_1/weatherapplication/Model/Current/Wind.java
@@ -1,22 +1,22 @@
 package com.geekbrains.android_1.weatherapplication.Model.Current;
 
 public class Wind {
-    private int speed;
-    private int deg;
+    private float speed;
+    private float deg;
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(float deg) {
         this.deg = deg;
     }
 }
